+++ aufgaben/06-test-doubles-invoice-processing/src/main/java/ex06/billingplatform/InvoiceResult.java
@@ -1,16 +1,16 @@
 package ex06.billingplatform;
 
-import java.math.BigDecimal;
-
 /**
  * Value type. Would normally be a Java record (see project README).
+ *
+ * <p>{@code total} is expressed in cents.
  */
 public class InvoiceResult {
 
     private String invoiceId;
-    private BigDecimal total;
+    private int total;
 
-    public InvoiceResult(String invoiceId, BigDecimal total) {
+    public InvoiceResult(String invoiceId, int total) {
         this.invoiceId = invoiceId;
         this.total = total;
     }
@@ -23,11 +23,11 @@ public class InvoiceResult {
         this.invoiceId = invoiceId;
     }
 
-    public BigDecimal getTotal() {
+    public int getTotal() {
         return total;
     }
 
-    public void setTotal(BigDecimal total) {
+    public void setTotal(int total) {
         this.total = total;
     }
